refactor: move resolveWikilink into @kolu/solid-markdown, beside its parser
The lens-debate had co-located resolveWikilink with resolveLinkHref in
@kolu/solid-browser, arguing both are "markdown reference → repo path." On
re-consulting /lowy and /hickey, hickey established they are two distinct
concepts (zero shared inputs/algorithm): resolveLinkHref is directory-relative
path math with no file list — a *browsing* concern that electricity.mdx already
assigns to solid-browser; resolveWikilink is wikilink-feature-specific vault
resolution (.md implied, alias/heading, ambiguity-surfacing) whose rules are
defined by the wikilink grammar and co-evolve with the markedWikilink parser.

A wikilink is a Markdown construct, so its resolver belongs in @kolu/solid-markdown
beside its parser — a pure, node-pure, host-called export (like url-policy's
safeHref), never wired into the render path, so the renderer keeps no file-list
dependency. solid-browser stays browsing-only; resolveLinkHref stays there.

Also folds in hickey's nit: the mutable `let cands` reassignment is now a single
`const` (base filter + conditional directory narrowing). Tests move to
solid-markdown/src/wikilink.test.ts. Atlas note + D2 updated to reflect the home.
