+++ packages/solid-browser/src/index.ts
@@ -8,10 +8,5 @@ export {
   type CreateBrowserOptions,
   DEFAULT_MAX_ENTRIES,
 } from "./createBrowser";
-export {
-  resolveLinkHref,
-  resolveRelativePath,
-  resolveWikilink,
-  type WikilinkResolution,
-} from "./relativePath";
+export { resolveLinkHref, resolveRelativePath } from "./relativePath";
 export { type PreviewPathCodec, pathFromPreviewPathname } from "./previewPath";
