+++ packages/solid-markdown/src/index.ts
@@ -1,1 +1,2 @@
 export { Markdown, type MarkdownVariant } from "./Markdown";
+export { resolveWikilink, type WikilinkResolution } from "./wikilink";
